Using general fix without hardcoded fields.
